Re-throw socket errors from stomp to work around a bizarre bug.
Stomp queues are subclasses of Thread that do the connect operation in
__init__. That means that if the connect fails, the resulting instance will be
half-baked. Luckily, you can only get to it through a traceback since the
half-baked instance was never assigned anywhere, but the mint traceback
handler just happens to have access to such a traceback, and the formatter
tries to get a representation of this half-baked class. The Thread class just
happens to define such a __repr__ method, which crashes on an assertion that
checks that the class was initialized.

Hilarity ensues.